Moving timestamp components to global components
